+++ packages/plugin-calls-ui/src/containers/IncomingCall.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { gql, useMutation } from '@apollo/client';
+import { gql, useMutation, useQuery } from '@apollo/client';
 
 import { Alert } from '@erxes/ui/src/utils';
 import { ICustomer } from '../types';
@@ -7,7 +7,7 @@ import IncomingCall from '../components/IncomingCall';
 import { __ } from '@erxes/ui/src/utils/core';
 import { callPropType } from '../lib/types';
 import { extractPhoneNumberFromCounterpart } from '../utils';
-import { mutations } from '../graphql';
+import { mutations, queries } from '../graphql';
 
 interface IProps {
   closeModal?: () => void;
@@ -18,6 +18,8 @@ interface IProps {
 
 const IncomingCallContainer = (props: IProps, context) => {
   const [customer, setCustomer] = useState<any>({} as ICustomer);
+  const [customers, setCustomers] = useState<any>([]);
+
   const [channels, setChannels] = useState<any>();
 
   const [hasMicrophone, setHasMicrophone] = useState(false);
@@ -36,6 +38,12 @@ const IncomingCallContainer = (props: IProps, context) => {
     callUserIntegrations?.[0]?.inboxId;
 
   const [createCustomerMutation] = useMutation(gql(mutations.customersAdd));
+  const { data, loading } = useQuery(gql(queries.callCustomerDetail), {
+    fetchPolicy: 'network-only',
+    variables: {
+      customerPhone: phoneNumber,
+    },
+  });
 
   useEffect(() => {
     navigator.mediaDevices
@@ -53,18 +61,28 @@ const IncomingCallContainer = (props: IProps, context) => {
 
         Alert.error(errorMessage);
       });
+  }, []);
 
+  useEffect(() => {
+    if (!loading && data?.callsCustomerDetail) {
+      setCustomer(data.callsCustomerDetail);
+    }
+  }, [loading, data]);
+
+  useEffect(() => {
     if (phoneNumber && call?.id) {
       createCustomerMutation({
         variables: {
           inboxIntegrationId: inboxId,
           primaryPhone: phoneNumber,
-          queueName: call.groupName || ''
+          queueName: call.groupName || '',
         },
       })
         .then(({ data }: any) => {
-          setCustomer(data.callAddCustomer?.customer);
-          setChannels(data.callAddCustomer?.channels);
+          if (data?.callAddCustomer?.customer) {
+            setCustomers([data.callAddCustomer.customer]);
+            setChannels(data.callAddCustomer.channels);
+          }
         })
         .catch((e) => {
           Alert.error(e.message);
@@ -74,6 +92,7 @@ const IncomingCallContainer = (props: IProps, context) => {
 
   return (
     <IncomingCall
+      customers={customers}
       customer={customer}
       channels={channels}
       hasMicrophone={hasMicrophone}