+++ packages/ui-tickets/src/boards/components/editForm/Sidebar.tsx
@@ -1,24 +1,22 @@
-import FormGroup from '@erxes/ui/src/components/form/Group';
-import ControlLabel from '@erxes/ui/src/components/form/Label';
-import SelectTeamMembers from '@erxes/ui/src/team/containers/SelectTeamMembers';
-import React from 'react';
-import { RightContent } from '../../styles/item';
-import { IItem, IOptions } from '../../types';
-import SidebarConformity from './SidebarConformity';
-import { __ } from '@erxes/ui/src/utils';
-import SelectNewBranches from '@erxes/ui/src/team/containers/SelectNewBranches';
+import FormGroup from "@erxes/ui/src/components/form/Group";
+import ControlLabel from "@erxes/ui/src/components/form/Label";
+import SelectTeamMembers from "@erxes/ui/src/team/containers/SelectTeamMembers";
+import React from "react";
+import { RightContent } from "../../styles/item";
+import { IItem, IOptions } from "../../types";
+import SidebarConformity from "./SidebarConformity";
+import { __ } from "@erxes/ui/src/utils";
+import SelectNewBranches from "@erxes/ui/src/team/containers/SelectNewBranches";
 
-import SelectDepartments from '@erxes/ui/src/team/containers/SelectDepartments';
-import { IUser } from '@erxes/ui/src/auth/types';
-import { pipeline } from 'stream';
-import { currentUser } from '@erxes/ui/src/auth/graphql';
+import SelectDepartments from "@erxes/ui/src/team/containers/SelectDepartments";
+import { IUser } from "@erxes/ui/src/auth/types";
 
 type Props = {
   item: IItem;
   relations: any;
   saveItem: (doc: { [key: string]: any }) => void;
   sidebar?: (
-    saveItem?: (doc: { [key: string]: any }) => void,
+    saveItem?: (doc: { [key: string]: any }) => void
   ) => React.ReactNode;
   options: IOptions;
   renderItems: () => React.ReactNode;
@@ -28,7 +26,7 @@ type Props = {
       status,
       timeSpent,
     }: { _id: string; status: string; timeSpent: number; startDate?: string },
-    callback?: () => void,
+    callback?: () => void
   ) => void;
   childrenSection: () => any;
   currentUser: IUser;
@@ -36,7 +34,7 @@ type Props = {
 
 class Sidebar extends React.Component<Props> {
   render() {
-    const { item, saveItem, sidebar, childrenSection } = this.props;
+    const { item, saveItem, sidebar, childrenSection, relations } = this.props;
     const userOnChange = (usrs) => saveItem({ assignedUserIds: usrs });
     const onChangeStructure = (values, name) => saveItem({ [name]: values });
     const assignedUserIds = (item.assignedUsers || []).map((user) => user._id);
@@ -47,7 +45,7 @@ class Sidebar extends React.Component<Props> {
       <RightContent>
         <FormGroup>
           <FormGroup>
-            <ControlLabel>{__('Branches')}</ControlLabel>
+            <ControlLabel>{__("Branches")}</ControlLabel>
             <SelectNewBranches
               name="branchIds"
               label="Choose branches"
@@ -55,7 +53,7 @@ class Sidebar extends React.Component<Props> {
               onSelect={onChangeStructure}
               filterParams={{
                 withoutUserFilter: true,
-                searchValue: 'search term',
+                searchValue: "search term",
               }}
             />
           </FormGroup>
@@ -73,18 +71,42 @@ class Sidebar extends React.Component<Props> {
         </FormGroup>
 
         <FormGroup>
-          <ControlLabel>{__('Departments')}</ControlLabel>
-          <SelectDepartments
-            name="departmentIds"
-            label="Choose departments"
-            onSelect={onChangeStructure}
-            initialValue={item?.departmentIds}
-          />
+          {relations.map((relation) => {
+            if (
+              relation.type === "departmentIds" ||
+              relation.relationType === "tickets:department"
+            ) {
+              return (
+                <React.Fragment key={relation._id}>
+                  <ControlLabel>{__("Departments")}</ControlLabel>
+                  <SelectDepartments
+                    name="departmentIds"
+                    label="Choose departments"
+                    onSelect={onChangeStructure}
+                    initialValue={item?.departmentIds}
+                  />
+                </React.Fragment>
+              );
+            }
+
+            return null;
+          })}
         </FormGroup>
+
         {sidebar && sidebar(saveItem)}
 
         <SidebarConformity {...this.props} />
-        {childrenSection()}
+        {relations
+          .filter(
+            (relation) =>
+              relation.type === "childrenIds" ||
+              relation.relationType === "tickets:children"
+          )
+          .map((relation) => (
+            <React.Fragment key={relation._id}>
+              {childrenSection()}
+            </React.Fragment>
+          ))}
       </RightContent>
     );
   }